requirements.txt
streamlit
st-supabase-connection==0.0.6

